// frontend/src/types/database.types.d.ts
export type Database = {
    __InternalSupabase: {
        PostgrestVersion: "13.0.5";
    };
    public: {
        Tables: {
            brands: {
                Row: {
                    created_at: string | null;
                    description: string | null;
                    id: string;
                    is_active: boolean | null;
                    logo_url: string | null;
                    name: string;
                    slug: string;
                    updated_at: string | null;
                };
                Insert: {
                    created_at?: string | null;
                    description?: string | null;
                    id?: string;
                    is_active?: boolean | null;
                    logo_url?: string | null;
                    name: string;
                    slug: string;
                    updated_at?: string | null;
                };
                Update: {
                    created_at?: string | null;
                    description?: string | null;
                    id?: string;
                    is_active?: boolean | null;
                    logo_url?: string | null;
                    name?: string;
                    slug?: string;
                    updated_at?: string | null;
                };
                Relationships: [];
            };
            categories: {
                Row: {
                    created_at: string | null;
                    display_order: number | null;
                    id: string;
                    image_url: string | null;
                    is_active: boolean | null;
                    name: string;
                    slug: string;
                    updated_at: string | null;
                };
                Insert: {
                    created_at?: string | null;
                    display_order?: number | null;
                    id?: string;
                    image_url?: string | null;
                    is_active?: boolean | null;
                    name: string;
                    slug: string;
                    updated_at?: string | null;
                };
                Update: {
                    created_at?: string | null;
                    display_order?: number | null;
                    id?: string;
                    image_url?: string | null;
                    is_active?: boolean | null;
                    name?: string;
                    slug?: string;
                    updated_at?: string | null;
                };
                Relationships: [];
            };
            colors: {
                Row: {
                    created_at: string | null;
                    hex_code: string | null;
                    id: string;
                    name: string;
                    slug: string;
                };
                Insert: {
                    created_at?: string | null;
                    hex_code?: string | null;
                    id?: string;
                    name: string;
                    slug: string;
                };
                Update: {
                    created_at?: string | null;
                    hex_code?: string | null;
                    id?: string;
                    name?: string;
                    slug?: string;
                };
                Relationships: [];
            };
            materials: {
                Row: {
                    created_at: string | null;
                    id: string;
                    name: string;
                    slug: string;
                };
                Insert: {
                    created_at?: string | null;
                    id?: string;
                    name: string;
                    slug: string;
                };
                Update: {
                    created_at?: string | null;
                    id?: string;
                    name?: string;
                    slug?: string;
                };
                Relationships: [];
            };
// ...
            products: {
                Row: {
                    activo: boolean | null;
                    brand_id: string | null;
                    color_id: string | null;
                    created_at: string | null;
                    descripcion: string | null;
                    destacado: boolean | null;
                    disponible: boolean | null;
                    en_oferta: boolean | null;
                    fecha_actualizacion: string | null;
                    fecha_creacion: string | null;
                    id: string;
                    material_id: string | null;
                    medidas: string | null;
                    metros_por_caja: number | null;
                    moneda: string | null;
                    nombre: string;
                    nuevo: boolean | null;
                    pei: string | null;
                    precio: number;
                    precio_anterior: number | null;
                    precio_metro: number | null;
                    sku: string;
                    slug: string;
                    stock: number | null;
                    unidad: string | null;
                    updated_at: string | null;
                };
                Insert: {
                    activo?: boolean | null;
                    brand_id?: string | null;
                    color_id?: string | null;
                    created_at?: string | null;
                    descripcion?: string | null;
                    destacado?: boolean | null;
                    disponible?: boolean | null;
                    en_oferta?: boolean | null;
                    fecha_actualizacion?: string | null;
                    fecha_creacion?: string | null;
                    id?: string;
                    material_id?: string | null;
                    medidas?: string | null;
                    metros_por_caja?: number | null;
                    moneda?: string | null;
                    nombre: string;
                    nuevo?: boolean | null;
                    pei?: string | null;
                    precio: number;
                    precio_anterior?: number | null;
                    precio_metro?: number | null;
                    sku: string;
                    slug: string;
                    stock?: number | null;
                    unidad?: string | null;
                    updated_at?: string | null;
                };
                Update: {
                    activo?: boolean | null;
                    brand_id?: string | null;
                    color_id?: string | null;
                    created_at?: string | null;
                    descripcion?: string | null;
                    destacado?: boolean | null;
                    disponible?: boolean | null;
                    en_oferta?: boolean | null;
                    fecha_actualizacion?: string | null;
                    fecha_creacion?: string | null;
                    id?: string;
                    material_id?: string | null;
                    medidas?: string | null;
                    metros_por_caja?: number | null;
                    moneda?: string | null;
                    nombre?: string;
                    nuevo?: boolean | null;
                    pei?: string | null;
                    precio?: number;
                    precio_anterior?: number | null;
                    precio_metro?: number | null;
                    sku?: string;
                    slug?: string;
                    stock?: number | null;
                    unidad?: string | null;
                    updated_at?: string | null;
                };
                Relationships: [
                    {
                        foreignKeyName: "products_brand_id_fkey";
                        columns: ["brand_id"];
                        isOneToOne: false;
                        referencedRelation: "brands";
                        referencedColumns: ["id"];
                    },
                    {
                        foreignKeyName: "products_color_id_fkey";
                        columns: ["color_id"];
                        isOneToOne: false;
                        referencedRelation: "colors";
                        referencedColumns: ["id"];
                    },
                    {
                        foreignKeyName: "products_material_id_fkey";
                        columns: ["material_id"];
                        isOneToOne: false;
                        referencedRelation: "materials";
                        referencedColumns: ["id"];
                    }
                ];
            };
            subcategories: {
                Row: {
                    category_id: string;
                    created_at: string | null;
                    description: string | null;
                    display_order: number | null;
                    id: string;
                    image_url: string | null;
                    is_active: boolean | null;
                    name: string;
                    slug: string;
                    updated_at: string | null;
                };
                Insert: {
                    category_id: string;
                    created_at?: string | null;
                    description?: string | null;
                    display_order?: number | null;
                    id?: string;
                    image_url?: string | null;
                    is_active?: boolean | null;
                    name: string;
                    slug: string;
                    updated_at?: string | null;
                };
                Update: {
                    category_id?: string;
                    created_at?: string | null;
                    description?: string | null;
                    display_order?: number | null;
                    id?: string;
                    image_url?: string | null;
                    is_active?: boolean | null;
                    name?: string;
                    slug?: string;
                    updated_at?: string | null;
                };
                Relationships: [
                    {
                        foreignKeyName: "subcategories_category_id_fkey";
                        columns: ["category_id"];
                        isOneToOne: false;
                        referencedRelation: "categories";
                        referencedColumns: ["id"];
                    }
                ];
            };
            tags: {
                Row: {
                    created_at: string | null;
                    id: string;
                    name: string;
                    slug: string;
                };
                Insert: {
                    created_at?: string | null;
                    id?: string;
                    name: string;
                    slug: string;
                };
                Update: {
                    created_at?: string | null;
                    id?: string;
                    name?: string;
                    slug?: string;
                };
                Relationships: [];
            };
        };
        Views: {
            [_ in never]: never;
        };
        Functions: {
            [_ in never]: never;
        };
        Enums: {
            [_ in never]: never;
        };
        CompositeTypes: {
            [_ in never]: never;
        };
    };
};